# backend/openapi.yaml
openapi: 3.0.3
info:
  title: EduVerify API
  version: "1.0.0"
  description: >
    API REST del backend de EduVerify. Todas las respuestas siguen el envelope
    { status: 'success', data } / { status: 'error', message }.
    Ver backend/README.md para setup y arquitectura.
servers:
  - url: http://localhost:3001
security: []
...
paths:
...
  /api/users/me:
    get:
      tags: [users]
      summary: Perfil del usuario autenticado
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: success }
                  data: { $ref: '#/components/schemas/User' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [users]
      summary: Actualizar nombre del usuario autenticado
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre: { type: string, minLength: 2, maxLength: 120 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/User' }
        "400":
          description: Nada que actualizar o body inválido
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /api/users/me/dark-mode:
    patch:
      tags: [users]
      summary: Actualizar preferencia de dark mode
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [dark_mode]
              properties:
                dark_mode: { type: boolean }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      dark_mode: { type: boolean }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /api/users/me/avatar:
    post:
      tags: [users]
      summary: Subir avatar
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [avatar]
              properties:
                avatar:
                  type: string
                  format: binary
                  description: JPEG/PNG/WebP, máx 2MB
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      avatar_url: { type: string, format: uri, nullable: true }
        "400":
          description: Sin archivo, tipo no permitido o supera el tamaño máximo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /api/users/me/banner:
    post:
      tags: [users]
      summary: Subir banner
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [banner]
              properties:
                banner:
                  type: string
                  format: binary
                  description: JPEG/PNG/WebP, máx 5MB
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      banner_url: { type: string, format: uri, nullable: true }
        "400":
          description: Sin archivo, tipo no permitido o supera el tamaño máximo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /api/users/{id}/profile:
    get:
      tags: [users]
      summary: Perfil público de un usuario (con conteo de videos y suscriptores)
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id: { type: integer }
                      nombre: { type: string }
                      rol: { type: string, enum: [estudiante, profesor, creador] }
                      avatar_url: { type: string, format: uri, nullable: true }
                      banner_url: { type: string, format: uri, nullable: true }
                      video_count: { type: integer }
                      subscriber_count: { type: integer }
        "404": { $ref: '#/components/responses/NotFound' }
  /api/users/{id}/videos:
    get:
      tags: [users]
      summary: Videos publicados por un usuario (paginado)
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VideoListResponse' }

  # ── VIDEOS ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Token obtenido de /api/auth/login, /registro o /google. Payload:
        { sub: userId, email, rol, premium }. Expira en 7 días.
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 20, maximum: 50 }
    UserId:
      name: id
      in: path
      required: true
      schema: { type: integer }
...
  responses:
...
    Unauthorized:
      description: Token ausente, inválido o expirado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    Categoria:
      type: string
      enum: [Programación, Ciberseguridad, Matemáticas, Electrónica, Arte]
    ErrorResponse:
      type: object
      properties:
        status: { type: string, example: error }
        message: { type: string }
...
    User:
      type: object
      properties:
        id: { type: integer }
        nombre: { type: string }
        email: { type: string, format: email }
        rol: { type: string, enum: [estudiante, profesor, creador] }
        premium: { type: boolean }
        dark_mode: { type: boolean }
        avatar_url: { type: string, format: uri, nullable: true }
        banner_url: { type: string, format: uri, nullable: true }
...
    Video:
      type: object
      properties:
        id: { type: integer }
        titulo: { type: string }
        descripcion: { type: string, nullable: true }
        url_video: { type: string, format: uri }
        categoria: { $ref: '#/components/schemas/Categoria' }
        tipo: { type: string, enum: [grabado, envivo] }
        es_premium: { type: boolean }
        vistas: { type: integer }
        duracion: { type: string }
        created_at: { type: string, format: date-time }
        autor: { type: string }
        autor_id: { type: integer }
        author_avatar_url: { type: string, format: uri, nullable: true }
    VideoListResponse:
      type: object
      properties:
        status: { type: string, example: success }
        data:
          type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Video' }
            total: { type: integer }
            page: { type: integer }
            limit: { type: integer }
...